Работает всё, кроме сетевого аниматора

+++ Assets/Scripts/IDamagable.cs
@@ -4,5 +4,5 @@ using UnityEngine;
 
 public interface IDamagable
 {
-    void Damage(float damage);
+    void Damage(int damage);
 }
